Login page, redux, comments
Implemented redux for error messages and login authentication, which still needs work. Login page was worked on and validation was implemented for the fields. The Login API was worked on but an error was occurring due to the 'use client' in the login page. Added a signup page, which is not as completed as the login page.