Login & Account
Worked further on the login/signup aspects, as well as logging out. Sign up isn't completed but will mirror login. I've also implemented local storage for staying signed in. The navbar also reflects that a user is logged in and the authSlice file has been updated to hold more user data, such as the name and userId.