Work on Forms
Made progress on skills to complete the UI to it and implemented untested code for saving it. Worked on reduce for skill slice. Fixed the navbar logout to redirect the user to the homepage, moved a redirect for users who are not logged in that was placed in the accordion UI to the account page. Made some updates to the saveData to take in the table name so that it can be used by all forms, same as for the receive data function.